+++ src/js/directives/_index.js
@@ -4,6 +4,7 @@ import Collapse from './collapse';
 import Confirm from './confirm';
 import Focus from './focus';
 import Modal from './modal';
+import SetValue from './set-value';
 import Trigger from './trigger';
 import Wrap from './wrap';
 
@@ -14,6 +15,7 @@ export default [
   Confirm,
   Focus,
   Modal,
+  SetValue,
   Trigger,
   Wrap,
 ];