+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+Add directive "set-value". #17
+
+### Changed
+- Improve lib/color.js #18
 
 ## [0.1.2] - 2017-07-23
 ### Added